pp: Enable more warnings, fix count in expanding parameter

+++ src/tasku_file.c
@@ -26,9 +26,9 @@ tacc_file_p tacc_open(char *name) {
     fseek(f, 0, SEEK_SET);
 
     out_file = tacc_malloc(sizeof(tacc_file_s));
-    src = tacc_malloc((file_sz + 1) * sizeof(char));
+    src = tacc_malloc(((size_t) file_sz + 1) * sizeof(char));
 
-    read_sz = fread(src, sizeof(char), file_sz, f);
+    read_sz = fread(src, sizeof(char), (size_t) file_sz, f);
     tacc_assert(read_sz == (size_t) file_sz,
                 "failed to read file %s: got %zu bytes with errno %d",
                 name,
@@ -40,7 +40,7 @@ tacc_file_p tacc_open(char *name) {
 
     out_file->tacc_file_src = src;
     out_file->tacc_file_name = name;
-    out_file->tacc_file_len = file_sz;
+    out_file->tacc_file_len = (size_t) file_sz;
 
     return out_file;
 }
